+++ legacy/Makefile
@@ -8,6 +8,7 @@ OBJS += entropy.o
 OBJS += flash.o
 OBJS += layout.o
 OBJS += oled.o
+OBJS += random_delays.o
 OBJS += rng.o
 
 ifneq ($(EMULATOR),1)
