+++ legacy/common.c
@@ -83,23 +83,6 @@ void __assert_func(const char *file, int line, const char *func,
 
 void hal_delay(uint32_t ms) { usbSleep(ms); }
 
-void wait_random(void) {
-  int wait = random32() & 0xff;
-  volatile int i = 0;
-  volatile int j = wait;
-  while (i < wait) {
-    if (i + j != wait) {
-      shutdown();
-    }
-    ++i;
-    --j;
-  }
-  // Double-check loop completion.
-  if (i != wait || j != 0) {
-    shutdown();
-  }
-}
-
 void drbg_init() {
   uint8_t entropy[48] = {0};
   random_buffer(entropy, sizeof(entropy));
