+++ legacy/webusb.c
@@ -18,7 +18,7 @@
 
 #include <string.h>
 
-#include "common.h"
+#include "random_delays.h"
 #include "usb21_standard.h"
 #include "util.h"
 #include "webusb.h"
